requirements.txt
alembic==0.9.6
astroid==1.5.3
click==6.7
coverage==4.4.2
Flask==0.12.2
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.0
Flask-Testing==0.6.2
gunicorn==19.7.1
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
nose==1.3.7
psycopg2==2.7.3.1
py==1.5.2
PyJWT==1.5.3
pylint==1.7.4
pytest==3.2.5
python-dateutil==2.6.1
python-editor==1.0.3
six==1.11.0
SQLAlchemy==1.1.14
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11